feat(eventemitter): add prependListener method + test
Add prependListener method on the EventEmitter class, which will check to see if the passed
`eventName` already exists in the events map, if so; unshift() the passed `listener` on the listener
array associated with the passed `eventName`. If the `eventName` does not exist, it will simply set
the key(`eventName`) and it's paired value(`listener`) to the events map.