refactor(eventemitter): restructure types and methods
Add listener type; an object which contains a listener `fn`, and once a boolean property that will
later be used to determine events that are called only once then removed. Update methods to
accomodate the new type structure for listeners. Abstract methods and add a private _addListener
method to handle additional events being added to the events property on the eventemitter class.
Update tests to accomodate new changes.